backport Apache Arrow 15 from tentacle ceph repo
Apache Arrow was basically never updated and we got a ancient version
of 6.5, which breaks compilation due to using c++11, which the abseil
library version from Bookworm doesn't support anymore.

Upstream upgraded it's arrow git submodule [0] in response to compilation
errors for the (future) Ceph 20 Tentacle release, but it's not (yet)
backported to Squid.

As we check in the processed source tarball we cannot just update a
git submodule here, but rather I copied over the src/arrow directory
from ceph tentacle branch.

[0]: https://github.com/ceph/ceph/pull/56585

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>