Revert "bump version to 19.2.3-pve2"
This reverts the following commit:
 - bac3136be81e96f5c5cf9201f013cd8496d890e9.
 - 0f31f0cc1b45d4b5d452e0f6e1a5432fdddd14a6.
 - c3dd8234ddf8c2a16e4e8de63e45dcf47bd53de4.

This is because while the patch for also removing the dh_installinit
is nice to have to avoid confusion potential, it was actually not
required, init.d based script really was gone already, Fabian just
checked out an old state as we now upload ceph also to the PVE repo
(to ensure linkage fro librados2 works) but I forgot to upload the
newer versions there too.

As the version was never uploaded anywhere and the debdiff showed that
it's (basically, minus some small noise in compiled binaries) the same
thing anyway.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>